% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{cluster_asv_dada2}
\alias{cluster_asv_dada2}
\title{Cluster ASVs in a sequence table into OTUs}
\usage{
cluster_asv_dada2(seqtab, similarity = 99, nproc = 4)
}
\arguments{
\item{seqtab}{A sequence table eg. generated by dada2.}

\item{similarity}{OTU clustering distance (\%).}

\item{nproc}{Number of processors.}
}
\value{
A clustered sequence table
}
\description{
Takes the ASV sequence table generated by DADA2 (seqtab) or phyloseq and
clusters OTUs.
}
\details{
Filler
}
\examples{
\dontrun{
clustered.seqtab <- cluster_asv_dada2(asv.seqtab)
}
}
