% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_clean_merge}
\alias{phyloseq_clean_merge}
\title{Cleanly merge (using sequence as ASV name) phyloseq objects in a list}
\usage{
phyloseq_clean_merge(phyloseq_list)
}
\arguments{
\item{ps_list}{A list of phyloseq objects}
}
\value{
A phyloseq object with all objects in the list merged
}
\description{
Assigns taxonomy to a phyloseq object using decipher, and returns the tax
table
}
\details{
A list containing only phyloseq objects can be cleanly merged,

including identical ASVs where the ASV name differs in the objects.

ASV merging is done by sequence
}
\examples{
\dontrun{
merged_phyloseq <- phyloseq_clean_merge(list_of_phyloseq_objects)
}
}
