% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_root_tree}
\alias{phyloseq_root_tree}
\title{Root Tree in phyloseq Object}
\usage{
phyloseq_root_tree(ps)
}
\arguments{
\item{ps}{A phyloseq object containing an unrooted tree}
}
\value{
The same phyloseq object with a rooted tree
}
\description{
Roots an unrooted tree in a phyloseq object
}
\details{
The tree is rooted by the longest terminal branch.
}
\examples{
\dontrun{
expt.rooted <- root_phyloseq_tree(expt.unrooted)
}
}
