% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_tax_dada2}
\alias{phyloseq_tax_dada2}
\title{Classify a phyloseq object with dada2}
\usage{
phyloseq_tax_dada2(ps, tax_file)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A phyloseq object with new tax assignments added to the tax table.
}
\description{
Assigns taxonomy to a phyloseq object using dada2 assignTaxonomy
}
\details{
The sequence table is assigned taxonomy using assignTaxonomy with

the greengenes or GTDB training set. (Ranks: D;P;C;O;F;G;S)
}
\examples{
\dontrun{
tax_table(phyloseq_project) <- phyloseq_tax_dada2(phyloseq_project, tax_file)
}
}
