% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_tax_decipher}
\alias{phyloseq_tax_decipher}
\title{Classify a phyloseq object with decipher}
\usage{
phyloseq_tax_decipher(ps, tax_file, nprocs = 8)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A phyloseq object with new tax assignments merged into the tax table.
}
\description{
Assigns taxonomy to a phyloseq object using decipher, and returns the tax table
}
\details{
The sequence table is assigned taxonomy using DECIPHER with the

greengenes or GTDB training set. (Ranks: D;P;C;O;F;G;S)
}
\examples{
\dontrun{
tax_table(phyloseq_project) <- phyloseq_decipher_tax(phyloseq_project, tax_file)
}
}
