% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_taxonomy_imputation}
\alias{phyloseq_taxonomy_imputation}
\title{Replace missing taxonomy in a phyloseq object}
\usage{
phyloseq_taxonomy_imputation(
  ps,
  unknown_taxon = "_unidentified",
  unknown_sp = " sp",
  make_unique = FALSE,
  addmaxrank = FALSE
)
}
\description{
Replace missing taxonomy in a phyloseq object
}
