% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_write_dataset_biom}
\alias{phyloseq_write_dataset_biom}
\title{Export phyloseq object to .biom file}
\usage{
phyloseq_write_dataset_biom(
  ps,
  filePATH,
  filePREFIX,
  writeFASTA = TRUE,
  rename = FALSE,
  useREFSEQ = TRUE
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{filePATH}{The path to write the files}

\item{filePREFIX}{The prefix for the output files}

\item{writeFASTA}{Write the fasta file}

\item{rename}{Rename the ASVs to ASV1, ASV2, etc.}

\item{useREFSEQ}{Use the refseq slot for sequence data}
}
\description{
Exports the sequence table and sequences from a phyloseq object
to a .biom file
}
\details{
A sequence table and fasta file are written to the file path
}
