% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.mate.R
\name{phyloseq_clean_merge}
\alias{phyloseq_clean_merge}
\title{Cleanly merge phyloseq objects with different ASV names}
\usage{
phyloseq_clean_merge(phyloseq_names)
}
\arguments{
\item{ps_list}{A vector of phyloseq objects}
}
\value{
A phyloseq object with all objects in the list merged
}
\description{
Merges phyloseq objects using sequence hashes to cleanly merge ASVs.
}
\details{
Phyloseq objects can be cleanly merged, including identical ASVs

where the ASV name differs in the objects.

ASV merging is done by sequence hash
}
\examples{
\dontrun{
merged_phyloseq <- phyloseq_clean_merge(vector_of_phyloseq_objects)
}
}
